% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_survey.R
\name{addStyleInBlocks}
\alias{addStyleInBlocks}
\title{addStyleInBlocks}
\usage{
addStyleInBlocks(wb, sheet, style, start_row, total_rows, n = 3, skip = 0)
}
\arguments{
\item{wb}{The workbook to add the style to}

\item{sheet}{The sheet to add the style to}

\item{style}{The style to add}

\item{rows}{The rows to add the style to}
}
\description{
This function adds a style to a workbook in blocks of n rows, skipping m rows
}
