% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_survey.R
\name{build_data_entry_segments}
\alias{build_data_entry_segments}
\title{build_data_entry_segments}
\usage{
build_data_entry_segments(species_list, poles, file_name)
}
\arguments{
\item{species_list}{The species list to create data entry segments for}

\item{poles}{The poles to create data entry segments for}

\item{file_name}{The file name to save the data entry segments to}
}
\value{
A data frame containing the data entry segments
}
\description{
This function creates data entry segments for the given species list and poles
}
