% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_survey.R
\name{build_data_sheets}
\alias{build_data_sheets}
\title{spring_survey_names}
\usage{
build_data_sheets(
  species_list,
  poles,
  file_name = "out/Plant Phenology Survey/Nuolja Transect Phenology Datasheets.xlsx"
)
}
\arguments{
\item{species_list}{The species list to create spring survey names for}

\item{poles}{The poles to create spring survey names for}

\item{file_name}{The file name to save the survey data sheets to}
}
\value{
A data frame containing the survey data sheets
}
\description{
This function creates spring survey names for the given species list and poles
}
