% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_survey.R
\name{build_species_list}
\alias{build_species_list}
\title{survey_tables}
\usage{
build_species_list(df)
}
\arguments{
\item{df}{The data frame to create survey tables for}
}
\value{
A data frame containing the survey tables
}
\description{
This function creates survey tables for the given data frame
}
