% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_survey.R
\name{build_spring_data_sheets}
\alias{build_spring_data_sheets}
\title{build_spring_data_sheets'}
\usage{
build_spring_data_sheets(
  species_list,
  poles,
  file_name = "out/Plant Phenology Survey/Nuolja Transect Phenology Datasheets.xlsx"
)
}
\arguments{
\item{species_list}{The species list to create spring data sheets for}

\item{poles}{The poles to create spring data sheets for}

\item{file_name}{The file name to save the spring data sheets to}
}
\value{
A data frame containing the spring data sheets
}
\description{
This function creates spring data sheets for the given species list and poles
}
