% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{buildCSV}
\alias{buildCSV}
\title{Build CSV}
\usage{
buildCSV(dataset, filename)
}
\arguments{
\item{dataset}{A data frame containing the dataset to process.}

\item{filename}{A character string representing the file name.}
}
\value{
A list containing the result of matrix calculation.
}
\description{
Build and write CSV files for subplots and plots.
}
