% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{closest}
\alias{closest}
\title{Closest}
\usage{
closest(e, df)
}
\arguments{
\item{e}{A numeric vector representing the point coordinates.}

\item{df}{A data frame containing the reference points.}
}
\value{
A numeric vector representing the section and distance.
}
\description{
Find the closest point in the dataframe to a given point, and return the corresponding section and distance.
}
