% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\docType{data}
\name{colnames}
\alias{colnames}
\title{Column Names}
\format{
An object of class \code{character} of length 10.
}
\usage{
colnames
}
\description{
A character vector containing the column names for the output data frame.
}
\keyword{datasets}
