% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{contemporary}
\alias{contemporary}
\title{Contemporary Filter}
\usage{
contemporary(z)
}
\arguments{
\item{z}{A vector of characters to filter.}
}
\value{
A character vector with contemporary perspective.
}
\description{
Filter input to identify contemporary perspective.
}
