% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{create_backup}
\alias{create_backup}
\title{Create a Backup of the Log File}
\usage{
create_backup(log_file)
}
\arguments{
\item{log_file}{A character vector specifying the name of the log file.}
}
\value{
NULL (This function is called for its side effect of creating a backup.)
}
\description{
This function creates a backup of the specified log file by renaming it with a
timestamp. If a backup file already exists, it is removed before creating the new
backup.
}
\examples{
create_backup("error.log")
}
