% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{create_file_structure}
\alias{create_file_structure}
\title{Create the File Structure for Logging}
\usage{
create_file_structure(log_file)
}
\arguments{
\item{log_file}{A character vector specifying the name of the log file.}
}
\value{
NULL (This function is called for its side effect of creating log files.)
}
\description{
This function creates a backup of the specified log file and initializes the
necessary log files for error and warning messages. It also creates a directory
for the log files if it does not already exist.
}
\examples{
create_file_structure("error.log")
}
