% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{createDateAndFix}
\alias{createDateAndFix}
\title{Create Date and Fix}
\usage{
createDateAndFix(string)
}
\arguments{
\item{string}{A character string representing a date.}
}
\value{
A Date object.
}
\description{
Convert a string to a Date object, fixing format if necessary.
}
