% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{createDir}
\alias{createDir}
\title{createDir}
\usage{
createDir(subdir)
}
\arguments{
\item{subdir}{A character string representing the name of the sub directory.}
}
\description{
Create a new sub directory inside the main path.
}
