% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_survey.R
\docType{data}
\name{DATA_FILE_PATTERN}
\alias{DATA_FILE_PATTERN}
\title{phenology_survey}
\format{
An object of class \code{character} of length 1.
}
\usage{
DATA_FILE_PATTERN
}
\description{
This script generates phenology survey datasheets and data entry segments based on species observation data.
}
\keyword{datasets}
