% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{dataframeBuilder}
\alias{dataframeBuilder}
\title{Build Data Frame}
\usage{
dataframeBuilder(data)
}
\arguments{
\item{data}{A list of data frames to accumulate.}
}
\value{
A data frame containing the accumulated data.
}
\description{
Build a data frame by accumulating rows from provided data.
}
