% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_survey.R
\docType{data}
\name{datasheet_info}
\alias{datasheet_info}
\title{Load required libraries}
\format{
An object of class \code{data.frame} with 176 rows and 3 columns.
}
\usage{
datasheet_info
}
\description{
Load the required libraries for the script'
}
\keyword{datasets}
