% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{delist}
\alias{delist}
\title{Delist}
\usage{
delist(y)
}
\arguments{
\item{y}{A list of character vectors.}
}
\value{
A character vector with concatenated elements.
}
\description{
Convert a list of character vectors into a single character vector with concatenated elements.
}
