% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{drawPlots}
\alias{drawPlots}
\title{Draw Plots}
\usage{
drawPlots(df)
}
\arguments{
\item{df}{A data frame containing the data to plot.}
}
\value{
None
}
\description{
Generate and save plots based on the provided data frame.
}
