% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{exportCSV}
\alias{exportCSV}
\title{exportCSV}
\usage{
exportCSV(filenames, filename, valid)
}
\arguments{
\item{filenames}{A character vector representing the paths to the input files.}

\item{filename}{A character string representing the name of the output file.}
}
\value{
A list containing the data frame and the output file name.
}
\description{
Export the data from the input files to a CSV file.
}
