% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{extract_date}
\alias{extract_date}
\title{Extract Date from Filename}
\usage{
extract_date(filename)
}
\arguments{
\item{filename}{A character string representing the filename.}
}
\value{
A character string representing the extracted date, or NULL if no date is found.
}
\description{
Extract the date from a filename in YYYYMMDD format.
}
