% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{formatDate}
\alias{formatDate}
\title{Format Date}
\usage{
formatDate(x)
}
\arguments{
\item{x}{A character string representing a date.}
}
\value{
A Date object.
}
\description{
Extract and format dates from a specific string format.
}
