% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{get_output_path}
\alias{get_output_path}
\title{get output path}
\usage{
get_output_path(path, filename)
}
\arguments{
\item{path}{The input path}
}
\value{
The output path
}
\description{
This function returns the output path for the given input path
}
