% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{getDataFilesPaths}
\alias{getDataFilesPaths}
\title{Get Data Files Paths}
\usage{
getDataFilesPaths(dir, pattern = "*.csv", full.names = TRUE)
}
\arguments{
\item{dir}{A character string representing the directory to search within.}

\item{pattern}{A character string representing the pattern to match. Default is "*.csv".}

\item{full.names}{A logical value indicating whether to return full names of the files. Default is TRUE.}
}
\value{
A character vector of file paths matching the pattern.
}
\description{
Retrieve the paths of data files within a given directory, optionally matching a pattern.
}
