% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{getDirs}
\alias{getDirs}
\title{Get Directories}
\usage{
getDirs(dir = "/data", pattern = "", full.names = FALSE)
}
\arguments{
\item{dir}{A character string representing the directory to search within. Default is "/data".}

\item{pattern}{A character string representing the pattern to match. Default is an empty string.}

\item{full.names}{A logical value indicating whether to return full names of the directories. Default is FALSE.}
}
\value{
A character vector of directory names matching the pattern.
}
\description{
Retrieve the names of directories within a given directory, optionally matching a pattern.
}
