% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{getFileNames}
\alias{getFileNames}
\title{Get File Names}
\usage{
getFileNames(dir = "/repack")
}
\arguments{
\item{dir}{A character string representing the directory to search within. Default is "/repack".}
}
\value{
A character vector of file names.
}
\description{
Retrieve the names of all Snow_Data_YYYY.csv files in a specified directory.
}
