% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getName}
\alias{getName}
\title{Get Name}
\usage{
getName(sourceFileName, specification)
}
\arguments{
\item{sourceFileName}{A character string representing the source file name.}

\item{specification}{A character string representing the specification to append to the file name.}
}
\value{
A character string representing the generated file name.
}
\description{
Generate a file name based on the source file name and a specification.
}
