% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{getPaths}
\alias{getPaths}
\title{Get Paths}
\usage{
getPaths(dir = "/data", pattern = "", full.names = TRUE)
}
\arguments{
\item{dir}{A character string representing the directory to search within. Default is "/data".}

\item{pattern}{A character string representing the pattern to match. Default is an empty string.}

\item{full.names}{A logical value indicating whether to return full names of the directories. Default is TRUE.}
}
\value{
A character vector of directory paths matching the pattern.
}
\description{
Retrieve the paths of directories within a given directory, optionally matching a pattern.
}
