% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{historical}
\alias{historical}
\title{Historical Filter}
\usage{
historical(y)
}
\arguments{
\item{y}{A vector of characters to filter.}
}
\value{
A character vector with historical perspective.
}
\description{
Filter input to identify historical perspective.
}
