% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\docType{data}
\name{keyset}
\alias{keyset}
\title{Keyset}
\format{
An object of class \code{list} of length 2.
}
\usage{
keyset
}
\description{
A list containing the keys for contemporary and historical data.
}
\keyword{datasets}
