% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{loadTransectDescription}
\alias{loadTransectDescription}
\title{Load Transect Description}
\usage{
loadTransectDescription(
  path = paste(getwd(), "/descriptions", sep = ""),
  pattern = "transect_description.csv"
)
}
\arguments{
\item{path}{A character string representing the path to the directory containing the transect description file. Default is the "descriptions" directory in the current working directory.}

\item{pattern}{A character string representing the pattern to match the transect description file. Default is "transect_description.csv".}
}
\value{
A data matrix containing the transect description.
}
\description{
Load the transect description from a CSV file.
}
