% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{logErrorCounts}
\alias{logErrorCounts}
\title{Count and Log Errors for Each File and Type of Error}
\usage{
logErrorCounts(error_list, file, count_log_file)
}
\arguments{
\item{error_list}{A list of error messages generated during validation.}

\item{file}{The name of the file being processed.}

\item{count_log_file}{(Optional) A character vector specifying a file where error
counts are logged. Default is "log/error_count_summary.txt".}
}
\value{
NULL (This function is called for its side effect of logging error counts.)
}
\description{
This function aggregates the count of each type of validation error
encountered in a file and logs the summary to a separate error count file.
}
\examples{
logErrorCounts(c("Error: Missing latitude format.", "Error: Missing latitude format."), "data.txt")

}
