% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{matrixCalc}
\alias{matrixCalc}
\title{Matrix Calculation}
\usage{
matrixCalc(entries, filename, data)
}
\arguments{
\item{entries}{A data frame containing the entries to process.}

\item{filename}{A character string representing the file name.}

\item{data}{Additional data required for processing.}
}
\value{
A list containing historical and contemporary matrices.
}
\description{
Process entries and generate historical and contemporary matrices.
}
