% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{noneNum}
\alias{noneNum}
\title{filter out none numerical used for coordinates who use the E & N notation.}
\usage{
noneNum(y)
}
\description{
filter out none numerical used for coordinates who use the E & N notation.
}
