% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\docType{data}
\name{pattern}
\alias{pattern}
\title{Validate a single line against a specific format.}
\format{
An object of class \code{character} of length 1.
}
\usage{
pattern
}
\arguments{
\item{line}{A character vector representing a single line to be validated.}
}
\value{
A logical value indicating whether the line matches the format pattern.
}
\description{
This function checks whether a given line matches a specific format pattern.
}
\examples{
validateLine("NS-20220510-001 68.37261219N 18.69783 1195.186 O")

}
\keyword{datasets}
