% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{percSeg}
\alias{percSeg}
\title{Percentage Segment}
\usage{
percSeg(ds)
}
\arguments{
\item{ds}{A vector of numeric values representing the segments.}
}
\value{
A numeric vector of percentage differences.
}
\description{
Calculate the percentage difference between consecutive segments.
}
