% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\docType{data}
\name{pheno_codes}
\alias{pheno_codes}
\title{phenology_codes}
\format{
An object of class \code{data.frame} with 23 rows and 1 columns.
}
\usage{
pheno_codes
}
\description{
This function reads in the phenology codes from the given path and directories
}
\keyword{datasets}
