% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{PhenologyValidator}
\alias{PhenologyValidator}
\title{Phenology Validator Class
inherits from R6Class}
\value{
a list with lenient and strict validation
}
\description{
This class is used to validate phenology data.
}
\examples{
phenology_validator <- PhenologyValidator$new()
phenology_validator$validate("species, date, subplot, code")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{LENIENT_PATTERN}}{Regular expression patterns for lenient validation}

\item{\code{STRUCT}}{Regular expression patterns for strict validation'}

\item{\code{PATTERNS}}{Regular expression patterns for validating individual fields}

\item{\code{ACCEPTABLE_CODES}}{Data frame of acceptable phenology codes and species}

\item{\code{PREVIOUS_SPECIES}}{List of previously observed species}

\item{\code{missing_species}}{List of species that were found in data but are not in the acceptable codes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PhenologyValidator-new}{\code{PhenologyValidator$new()}}
\item \href{#method-PhenologyValidator-validate}{\code{PhenologyValidator$validate()}}
\item \href{#method-PhenologyValidator-validateField}{\code{PhenologyValidator$validateField()}}
\item \href{#method-PhenologyValidator-clone}{\code{PhenologyValidator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenologyValidator-new"></a>}}
\if{latex}{\out{\hypertarget{method-PhenologyValidator-new}{}}}
\subsection{Method \code{new()}}{
Regular expression patterns for strict validation


Regular expression patterns for strict validation


Regular expression patterns for validating individual fields


Data frame of acceptable phenology codes and species


List of previously observed species


Initialize the PhenologyValidator class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenologyValidator$new(
  csv_path =
    "descriptions/Nuolja Master Documents/Accepted_PhenoCodes_Species_CURRENT.csv"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{csv_path}}{Path to the CSV file containing acceptable phenology codes and species}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenologyValidator-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PhenologyValidator-validate}{}}}
\subsection{Method \code{validate()}}{
Validate a line of phenology data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenologyValidator$validate(line)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{line}}{A string representing a line of phenology data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with lenient validation results and field-wise validation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenologyValidator-validateField"></a>}}
\if{latex}{\out{\hypertarget{method-PhenologyValidator-validateField}{}}}
\subsection{Method \code{validateField()}}{
List of species that were found in data but are not in the acceptable codes


Validate individual fields in a line of phenology data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenologyValidator$validateField(strLine)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strLine}}{A string representing a line of phenology data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list indicating the validity of each field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenologyValidator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PhenologyValidator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenologyValidator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
