% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\docType{data}
\name{plots_and_subplots}
\alias{plots_and_subplots}
\title{read_descriptions'}
\format{
An object of class \code{data.frame} with 78 rows and 3 columns.
}
\usage{
plots_and_subplots
}
\description{
columns Poles, Plot and Subplot
}
\keyword{datasets}
