% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{printValidationError}
\alias{printValidationError}
\title{Validate Line and Print Field-Specific Errors}
\usage{
printValidationError(
  line,
  validation,
  file = NA,
  line_number = NA,
  log_file = NULL
)
}
\arguments{
\item{line}{A character vector representing the line to be validated.}

\item{file}{(Optional) The name of the file containing the line (default is NA).}

\item{line_number}{(Optional) The line number within the file (default is NA).}

\item{log_file}{(Optional) A character vector specifying a file where error messages
are logged. If NULL, messages are only printed to the console.}
}
\value{
NULL (This function is called for its side effect of printing/logging messages.)
}
\description{
This function validates a line against a specified format pattern and prints/logs
detailed errors for each field that does not match the expected format.
}
\examples{
printValidationError ("NS-20220510-001 68.37261219N 18.69783 1195.186 O", "data.txt", 5, "log/validation.txt", "log/warnings.txt")

}
