% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{process_phenology_data}
\alias{process_phenology_data}
\title{process_phenology_data}
\usage{
process_phenology_data(input, dirs)
}
\arguments{
\item{input}{The input data frame}

\item{dirs}{The directories to search for the phenology data}
}
\description{
This function processes the phenology data from the given input and directories
}
