% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{process_snow_data}
\alias{process_snow_data}
\title{Process Snow Data}
\usage{
process_snow_data(paths, dirs)
}
\arguments{
\item{dirs}{A character vector of directory names.}

\item{path}{A character string representing the path to the directory.}
}
\value{
A list containing the processed data.
}
\description{
Process Snow_Data_YYYY.csv files in a specified directory.
}
