% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{projection}
\alias{projection}
\title{Projection}
\usage{
projection(transect_desc, x)
}
\arguments{
\item{transect_desc}{A data frame containing the transect description.}

\item{x}{A numeric vector representing the point coordinates.}
}
\value{
A numeric vector representing the distance along the transect and the perpendicular distance to the line.
}
\description{
Calculate the projection of a point onto the transect line.
}
