% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{read_phenology_data}
\alias{read_phenology_data}
\title{process_phenology_data}
\usage{
read_phenology_data(dir, all = FALSE)
}
\arguments{
\item{path}{The path to the directory containing the phenology data}

\item{dirs}{The directories to search for the phenology data}
}
\value{
A data frame containing the phenology data
}
\description{
This function reads in all phenology data from the given path and directories and processes it into two files; Observation by Year, Species, Code and Subplot; Second First observation by Year, Species and Code
}
