% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{readFile}
\alias{readFile}
\title{Read File}
\usage{
readFile(x, valid)
}
\arguments{
\item{x}{A character string representing the file path.}
}
\value{
A processed data frame.
}
\description{
Read and process a data file.
}
