% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{section}
\alias{section}
\title{Section}
\usage{
section(df, x)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{x}{A numeric value representing the projection factor.}
}
\value{
A numeric vector representing the section of the dataframe.
}
\description{
Find the section of the dataframe corresponding to a given projection factor.
}
