% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{SnowValidator}
\alias{SnowValidator}
\title{Snow Validator Class'
inherits from R6Class}
\value{
a list with lenient and strict validation
}
\description{
This class is used to validate snow data.
}
\examples{
snow_validator <- SnowValidator$new()
snow_validator$validate("NS-20230101-001 60.123456N 15.123456E 100.0 os")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{STRICT_PATTERN}}{Regular expression patterns for strict validation'}

\item{\code{LENIENT_PATTERN}}{Regular expression patterns for lenient validation}

\item{\code{PATTERNS}}{Regular expression patterns for validating individual fields}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SnowValidator-new}{\code{SnowValidator$new()}}
\item \href{#method-SnowValidator-validate}{\code{SnowValidator$validate()}}
\item \href{#method-SnowValidator-validateField}{\code{SnowValidator$validateField()}}
\item \href{#method-SnowValidator-clone}{\code{SnowValidator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnowValidator-new"></a>}}
\if{latex}{\out{\hypertarget{method-SnowValidator-new}{}}}
\subsection{Method \code{new()}}{
Regular expression patterns for strict validation


Regular expression patterns for strict validation


Regular expression patterns for validating individual fields


Initialize the SnowValidator class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowValidator$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnowValidator-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SnowValidator-validate}{}}}
\subsection{Method \code{validate()}}{
Validate a line of snow data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowValidator$validate(line)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{line}}{A string representing a line of snow data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with lenient and strict validation results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnowValidator-validateField"></a>}}
\if{latex}{\out{\hypertarget{method-SnowValidator-validateField}{}}}
\subsection{Method \code{validateField()}}{
Validate individual fields in a line of snow data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowValidator$validateField(line)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{line}}{A string representing a line of snow data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list indicating the validity of each field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnowValidator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SnowValidator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowValidator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
