% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_survey.R
\name{spring_survey_names}
\alias{spring_survey_names}
\title{spring_survey_names}
\usage{
spring_survey_names(species_list, poles, i)
}
\arguments{
\item{species_list}{The species list to create spring survey names for}

\item{poles}{The poles to create spring survey names for}

\item{i}{The index of the poles to create spring survey names for}
}
\value{
A data frame containing the spring survey names
}
\description{
This function creates spring survey names for the given species list and poles
}
