% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{subCalc}
\alias{subCalc}
\title{Subplot Calculation}
\usage{
subCalc(entries, filename, vs, data)
}
\arguments{
\item{entries}{A data frame containing the entries to process.}

\item{filename}{A character string representing the file name.}

\item{vs}{A data frame containing additional variables for processing.}

\item{data}{Additional data required for processing.}
}
\value{
A list containing contemporary and historical subplot data.
}
\description{
Calculate and generate contemporary and historical subplots.
}
