% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{sumPerc}
\alias{sumPerc}
\title{Sum Percentage}
\usage{
sumPerc(m)
}
\arguments{
\item{m}{A data matrix containing percentage values.}
}
\value{
A list with contemporary, historical, and default percentage values.
}
\description{
Calculate the sum of percentages for contemporary and historical data.
}
