% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_survey.R
\name{survey_data_sheet_get}
\alias{survey_data_sheet_get}
\title{survey_data_sheet_get}
\usage{
survey_data_sheet_get(species_list, poles, i)
}
\arguments{
\item{species_list}{The species list to create survey data sheets for}

\item{poles}{The poles to create survey data sheets for}

\item{i}{The index of the poles to create survey data sheets for}
}
\value{
A data frame containing the survey data sheets
}
\description{
This function creates survey data sheets for the given species list and poles
}
