% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{write}
\alias{write}
\title{Write Data}
\usage{
write(data, file, t)
}
\arguments{
\item{data}{A data frame containing the data to write.}

\item{file}{A character string representing the file name.}

\item{t}{A character string representing the type of data.}
}
\value{
None
}
\description{
Write data to a CSV file.
}
