% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{formatDate}
\alias{formatDate}
\title{function that destill dates from format type NS-20190505-033 and NS20200805-033 and return the date in the format 2020-05-05}
\usage{
formatDate(x)
}
\description{
function that destill dates from format type NS-20190505-033 and NS20200805-033 and return the date in the format 2020-05-05
}
