% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{noneNum}
\alias{noneNum}
\title{Filter Non-numerical Characters}
\usage{
noneNum(y)
}
\arguments{
\item{y}{A vector of characters to filter.}
}
\value{
A vector of double values after removing non-numerical characters.
}
\description{
Filter out non-numerical characters from input.
}
