% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{printValidationError}
\alias{printValidationError}
\title{Print validation error messages for lines that do not match the format.}
\usage{
printValidationError(line, file, line_number)
}
\arguments{
\item{line}{A character vector representing a single line to be validated.}
}
\description{
This function prints error messages for lines that do not match the expected format.
}
\examples{
printValidationError("NS-20220510-001 68.37261219N 18.69783 1195.186 X")

}
